Fixed up Graveyard
Level 2 is now a bit smaller and added a path in the middle of the graveyard while also removing the double path at the beginning. Cleaned up any gaps that were seeable and fixed some gates that were in the path and added some to empty parts of the fence.